requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=0.19.0,<0.20.0
requests>=2.26.0,<2.27.0
